project.py: fix windows-style paths as manifest URLs
If the manifest URL is a windows-style path,
e.g. C:\Users\marti\my-manifest, then the current code produces a
manifest.path configuration option set to '\Users\marti\my-manifest'
instead of the expected 'my-manifest'.

Fix it by using PurePath so we get PureWindowsPath on windows, instead
of using the low-level posixpath. Note that PurePath is PurePosixPath
on POSIX platforms, so the behavior is the same on those.

Fix the import order while we are here, moving stdlib imports to their
proper section.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>