tests: relax check_proj_consistency url checks on windows
The test_import_project*() cases in test_project.py are failing on
windows for reasons that are irrelevant and ultimately down to path
comparisons not being canonicalized for that platform.

Rather than fix tests I don't really think are broken, check if the
expected URL is an actual directory and fall back on PurePath
comparisons instead.

Keep the behavior the same for macOS and Linux.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>