github: Fix workflow to run on differents host (mac, win)
Fix copy/paste issue in which the 'runs-on' was always ubuntu-latest so
we didn't actually do anything with the matrix.os to run on mac or
windows.

Also cleanup some minor formatting and report the sys.platform so we can
verify in the future we are on the correct platform.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>