manifest: type check 'import-context' kwarg
This is used internally to pass along an import context in a way that
can't casually be done by a user, since '-' is not a valid character
in an identifier.

Adding asserts about its value won't hurt, and will help convince the
type checker that this is doing the right thing later on.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>