manifest: accept os.PathLike, produce str
Modify the APIs so that most any PathLike object can be passed to
public constructors or methods, and the results behave correctly.

Preserve existing Project and Manifest attribute and property
semantics, though: attributes that were str are still str.

We also keep the second argument to ImporterType a str, since
there's no reason not to and it will avoid potentially requiring
future users to call os.fspath() on the argument just to be safe.

This allows us to avoid using west.util.canon_path(), especially to do
path comparisons in Manifest.get_projects(). Instead, let the standard
library do the right thing for us.

Along the way we find another way in which a ManifestProject is not
a Project and should not have been subclassed from it, because we are
forced to ignore type checking for the ManifestProject _path
attribute, since it has the wrong type compared to its parent class.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>