project.py: stop using canon_path()
Use pathlib.Path instead. To try to avoid mistakes, type-annotate some of
the auxiliary methods involved in this change to let mypy catch
situations where we're passing a Path to something that expects a str.
Don't go crazy with that, though; this is application code, not API,
so we don't need to get all Java in here.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>