remove .shippable.yml
Experience has borne out the hope that the CI system using GitHub
actions offers the same amount of testing on Linux platforms as
existing shippable infrastructure. It's also configured to run on
macOS and Windows, which is nicer. So there's no need to continue to
run shippable based testing now.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>