app: fix import error followed by update
Consider a manifest file which uses 'import' on a project, where that
project revision's manifest file contains invalid manifest data.
The manifest-rev branch in that project will be left pointing at the
bogus manifest. The only way out is to update the main west.yml and
point at a new revision, then run west update.

However, main.py refuses to run west update in exactly this situation,
leading to an unusable workspace.

Fix this and add a regression test.

This new ability does open us up to running "west update" even with an
invalid manifest and no imports. We handle this in main.py.

Reported-by: Pete Skeggs <peter.skeggs@nordicsemi.no>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>