tests: extended test case for Manifest.from_file()
Manifest.from_file() and Manifest.from_file(topdir=<topdir>) produces
different results.

The commit extends the test case to verify that creation of Manifest
object is identical when produced with topdir argument or without, as
long as the topdir provided as argument is identical to what
`util.west_topdir()` would have returned.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>