manifest: adjust some implementation details
It's going to be convenient to call _load_imap() from _load_project()
in order to get the path-prefix out of an "import:" block before we
have a Project instantiated. To do that we will need a way to call it
without the project instance being created.

Make it so with a small refactor.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>