manifest: add 'path-prefix' key to import: map
Several users have requested the ability to place a project and its
imports in a designated subdirectory to keep third party code
separated from user-written code. Add this feature using a new key in
an import: map.

Example syntax:

  manifest:
    projects:
    - name: zephyr
      url: https://github.com/zephyrproject-rtos/zephyr
      import:
        path-prefix: zephyr_projects

Semantics:

 - the zephyr project, and all imported projects, have zephyr_projects
   as an additional path component. e.g. zephyr_projects/zephyr instead
   of zephyr as the path for zephyr project

 - the path-prefix value cannot escape the workspace, e.g.
   path-prefix: ../.. would be an error in the above example.

 - the 'path-prefix' value is "cumulative" across imports, e.g.
   if the imported manifest in turn does an import with "path-prefix: foo",
   the combined directory is "zephyr_projects/foo"

We've already laid most of the groundwork for plumbing this through
the import handling code using the _import_ctx and _import_map
namedtuples. When recursively loading Manifest instances during import
resolution, _import_ctx is a container that allows us to pass
information obtained from "higher" manifests down to "lower" manifests
in the tree. The _import_map namedtuple is just a container for the
import: contents when its value is a dict.

What remains to be done is then to add the current "cumulative" path
prefix to _import_ctx and pass it around to a couple of additional
places where it wasn't previously needed but now is. This requires
generalizing _new_ctx() a bit so we can combine an existing context
with the entire _import_map at a particular "import:" instead of just
using its filter function like we do now.

The one edge case is that since path-prefix affects the "imported"
project itself (zephyr in the above example), we need to make sure to
load it when creating every Project instance. This means we
effectively do it twice, which is a bit wasteful but keeps the code
simple, so it's worthwhile waste.

Fixes: #367
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>