tests: clean up dependency handling
Recent versions of tox are complaining with errors that look like
this:

  WARNING: test command found but not installed in testenv
    cmd: C:\Users\Marti\Envs\west-dev\Scripts\mypy.EXE
    env: C:\Users\Marti\src\west\.tox\py3-windows
  Maybe you forgot to specify a dependency? See also the allowlist_externals envconfig setting.

  DEPRECATION WARNING: this will be an error in tox 4 and above!

Fix that by running all the dependencies through the python binary in
the tox virtualenv. Moving the dependencies out of tox.ini seems to be
necessary too; I couldn't get this to work otherwise. It's more in
keeping with the style in the docs here anyway:

https://tox.readthedocs.io/en/latest/config.html

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>