Improve west init -l docs
Let's make it clearer what happens in this case.

For the sake of not making the command line help too long, let's drop
the point that it can't be combined with -m or --mr and just make it
clear that -l doesn't use MANIFEST_URL. People will find out right
away if they try it that it doesn't work if they combine these
options.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>