manifest: rework workspace escape checks
Let's use a lexical check instead of escapes_directory(). This
continues to let some users who are symlinking outside the workspace
to keep doing that, even though we don't officially support it in
west.

Also throw in a check for an absolute project path, which should never
happen and is similarly a sign of an attempt to escape the workspace.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>