manifest: add project group support
We currently don't have a way to group projects with extra semantic
information. There is a use case for this in Zephyr; the ultimate goal
is to allow manifests to not update certain groups of projects by
default, but allow easy update of those groups if the user opts in.

Start to enable this by letting a 'projects' item in a manifest file
have a 'groups' key, which must contain a list of scalar values (str,
int, or float).

For example, in this manifest:

  manifest:
    # ...
    projects:
    - name: foo
      groups:
      - YouAreIt
    - name: bar
      groups:
      - YouAreIt
      - MeToo
    - name: baz

The projects have these groups lists:

- foo: ['YouAreIt']
- bar: ['YouAreIt', 'MeToo']
- baz: []

As a restriction to give us some room to grow into and generally keep
things sane, 'projects' group values must not contain commas (,),
colons (:), whitespace. Groups cannot start with a '-' either as
that's how we'll be blocking them.

The manifest data are converted to lists of str internally and
associated with the west.manifest.Project objects.

Projects are "active" when any of their groups are not blocked.

Groups are allowed by default, but may be blocked using the manifest,
like this:

  manifest:
    ...
    groups: [-thisisblocked,-this-too]

Or via the .west/config workspace local configuration file, like this:

  [manifest]
  groups = -blocked1,-blocked2,allowed

I.e. via commands like:

  $ west config manifest.groups -- -blocked1,-blocked2,allowed

There is also an API routine is_active() for checking if a project is
active, potentially with some extra blocklist and allowlist entries
for groups obtained from elsewhere.

For now, no commands are aware of this. We'll make commands respond
differently to 'inactive' projects in the next patches.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>