west list: teach it about project groups
Add '-f {groups}' and '--all' options. Respect active projects by not
printing inactive ones by default.

This allows printing a project's groups, and hides non-updatable
groups by default.

Here we rely on the fact that groups may not have commas or whitespace
to ensure that the printed-out groups are a single whitespace-free and
comma separated string in the 'west list' output.

For example, in this manifest:

  manifest:
    # ...
    projects:
    - name: foo
      groups:
      - foo-group-1
      - foo-group-2
    - name: bar
      path: path-for-bar
    - name: baz
      groups:
      - baz-group

The output for "west list -f '{name} {groups} {path}'" is the following,
with <SPC> standing for a space character:

    foo<SPC>foo-group-1,foo-group-2<SPC>foo
    bar<SPC><SPC>path-for-bar
    baz<SPC>baz-group<SPC>baz

The empty string in the {groups} field for 'bar' can be picked up by
tools like cut(1) to find the entire list of groups. E.g., this command

    $ west list -f '{name} {groups} {path} | cut -d' ' -f 2

will print:

    foo-group-1,foo-group-2

    baz-group

Additionally, we repurpose the old --all command line option (that used
to print the location of the bootstrapped west repository back in the
0.5.x days) to print projects that would be skipped by 'west update'
by default due to groups or no-groups lists.

We're about to do something similar to several other commands, so
structure the help in the epilog in a way that we'll take advantage of
in subsequent patches.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>