west update: teach it about project groups
Add support in "west update" for skipping updates for any project that
the Manifest object says is not active.

In order to allow for overrides, let the user specify additional
groups to add to the allowlist via a new "--group" argument, and
additional groups to add to the blocklist via "--no-group". This is
useful for debugging credential management and doing things like
quickly syncing the entire potential workspace with a command like
"west update --groups all".

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>