manifest: fix up and rename a submodules helper
The _get_submodules() method is setting a 'submodules' attribute on
the Manifest object. It should just return a value instead. The
submodules data is per-project, so there's no need for the Manifest
object to track this, since it already has a list of projects.

Fix that, renaming the method to _load_submodules() as well for
consistency with the other helper procedures called by _load().

Move its definition underneath _validate_project_groups() while we're
here, since that is the previous step inside the caller
(_load_project()).

Align the error handling messages with _load_imap()'s by printing the
name of the project that's causing problems on error.

Convert the docstring to a comment like we do for all the other
private helpers, so west API users calling help() aren't encouraged to
rely on its behavior.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>