__init__: add null handler for 'west' logger
API modules like west.manifest are using the standard logging module.
This module prints messages at warning severity or above to stderr by
default.

That prints messages twice when a west command configures its own
logger, like this:

  $ west foo
  some warning message
  WARNING: some warning message

The first line comes from the print to stderr done by the default
handler. The second one is from a ProjectCommandLogHandler.

Disable the first message, which we do not want, by registering a null
handler for the top level. This results in just the
ProjectCommandLogHandler output being shown, like this:

  $ west foo
  WARNING: some warning message

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>