manifest: move some type aliases around
Move RawGroupType right before its only point of use.

Move GroupFilterType and GroupsType to the top level list of type
aliases, as prep work for using them in more places. Change the '#:'
comment style to '#' while we're here; the '#:' style is reserved for
comments which are also docstrings that end up in Sphinx, which
doesn't apply here.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>