update: move helper to staticmethod
This moves the definition closer to the point of use, allows us to
move the --stats bookkeeping into the method itself, and makes it more
obvious that 'update' is the only command that ought to be hitting the
network to fetch revisions.

For now, we don't need any instance data, but that will change in a
subsequent patch.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>