update: clean up setting new manifest-rev
Updating manifest-rev is a separate operation from fetching, but the
two are tracked and printed together in the --stats output.

Now that this code is in an instance method that has access to the
stats dict, it is easy to track manifest-rev and fetching separately,
so make it happen. Also resolve manifest-rev to a SHA before calling
_update_manifest_rev(), so the reflog entry says:

    west update: moving to <SHA>

instead of something like:

    west update: moving to FETCH_HEAD^{commit}

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>