configuration: fix WEST_CONFIG_LOCAL precedence
The documentation says that the WEST_CONFIG_LOCAL environment variable
must override the default file location, <topdir>/.west/config, if
set. It doesn't, though; fix that.

This requires us to be a bit more careful in how we write the fixture
for the west.configuration test suite, and requires updating a test
case to match the documentation.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>