tests: adjust create_repo() to keep working on OLDER git
Commit 6d42388f20959723eae1781537c7090a9600599d
("tests: adjust create_repo() to keep working on newer git") got tox
working again in Arch on my laptop, and in CI for whatever GitHub
decides to use for ubuntu-latest, but it breaks tox on my Ubuntu 20.04
LTS testing environment that I use for release testing.

Fix it, again, and hopefully for the last time. See comments in the
source code for details on the fix.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>