project: overwrite stdout/stderr objects as well
the existing objects might not lke the file descriptor of the log file.
Specifically the one of Windows' `cmd`.

So in addition to changing the fd we're now replacing the IO objects as well.
the mode got changed from `wb` to `w` because pythons stdout/stderr expects
it that way. Since `b` should not have any meaning outside of python
(for the fd) it shouldn't matter for the fd we're passing to dup2.
Opening the file multiple times would introduce a ton of issues - even with
`a`.

Fixes: #527
Signed-off-by: Michael Zimmermann <sigmaepsilon92@gmail.com>