update: add --fetch=tags option
Like '--fetch=smart', this makes 'west update' skip fetches for SHAs
that are already available locally. But unlike '--fetch=smart', it
forces any tags to get fetched again.

Although tags "shouldn't" change, in practice they do sometimes, and
this makes it a little faster to update the workspace in a way that
doesn't re-fetch SHAs unnecessarily.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>