.github: remove py 3.6 from ci matrix
We no longer have access to 3.6 in the system that provides our python
interpreter, so we'll have to do without it in CI.

I think in practice this is going to mean that 3.7 is the lowest
version of Python that west supports, but that's really a Zephyr
decision and I don't feel comfortable make it unilaterally here. So
for now we'll continue to treat failures on 3.6 as bugs, but we won't
test on that platform in CI. That's not ideal and I'll try to do
release testing on 3.6 for as long as I can, as it's the best we've
got.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>