WestCommand: use Configuration objects
Add a 'config' property to WestCommand. This is a
west.configuration.Configuration accessible for the duration of
do_run(). For the same reason as the 'WestCommand.manifest' property
exists the way it does, we cannot add a new constructor argument to
capture the configuration. Handle this as a kwarg to run().

Take a Configuration in extension_commands().

Handle the necessary changes in main.py. These changes should be
invisible to all the extension commands I am aware of in the wild.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>